requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
matplotlib>=3.3.0
scikit-image>=0.17.0
pillow>=8.0.0
tqdm>=4.50.0